chore: moved masking utils to utilities and mask_bird to root scripts
