+++ server/utilities/masking/linking_utils/corenlp.py
@@ -33,7 +33,7 @@ class CoreNLP:
             os.environ['CORENLP_HOME'] = os.path.abspath(
                 os.path.join(
                     os.path.dirname(__file__),
-                    '../../third_party/stanford-corenlp-full-2018-10-05'))
+                    '../../../scripts/third_party/stanford-corenlp-full-2018-10-05/'))
         if not os.path.exists(os.environ['CORENLP_HOME']):
             raise Exception(
                 f'''Please install Stanford CoreNLP and put it at {os.environ['CORENLP_HOME']}.
